feat(cluste-init): add onboard cmd
